Extract item tag commands
